Re-org, split out annotate and call_genes
there is a new command called call_genes or just call so that the user
can call genes this will then be part of the pipeline methodology. What
we now have is a complex click app that passes context. However, we need
to encapsulate the commands and we need to add the clean up at the end or
an explicit non clean up. This could be step by step. It may be necessary
to enforce the cording of some commands but I don't think that is the
case. It is yet to be seen if the commands can work well together.